Rework of the queries, so they are more conformant to Scientific Data Model phylosophy.
Deep OEB Scientific data curation has uncovered this tool was not
properly fetching the participant tools (it worked because the datasets were wrongly pointing to the right tool).

graphQL queries have been expanded, so it get both metric events, participant and assessment datasets associated to each challenge. And the code to process
it has been written.

The code (purposely) fails when the provided authorization does not give access to TestActions.